+++ lib/utils/format.dart
@@ -221,7 +221,7 @@ class Fmt {
   }
 
   static bool checkPassword(String pass) {
-    var reg = RegExp(r'^(?![0-9]+$)(?![a-zA-Z]+$)[\S]{6,20}$');
+    var reg = RegExp(r'^(?![0-9]+$)(?![a-zA-Z]+$)[\S]{8,20}$');
     return reg.hasMatch(pass);
   }